% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_raincloud.R
\name{geom_raincloud}
\alias{geom_raincloud}
\title{Raincloud plot with stats}
\usage{
geom_raincloud(
  mapping = NULL,
  color_fill = "lightgreen",
  side = c("left", "right"),
  violin_linewidth = 0,
  violin_alpha = 1,
  nudge_value = 0.05,
  dotsize = 1,
  stackratio = 1.4,
  binwidth = 0.5,
  quartile_linewidth = 1,
  fun = median,
  marker_size = 0.3
)
}
\value{
A ggplot2 layer
}
\description{
This function creates a custom raincloud plot, which is a combination of a half violin plot and a dotplot.
It is useful for visualizing the distribution of data with both density (violin) and individual data points (dotplot).
}
