% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_glyptotek.R
\name{glyptotek_colors}
\alias{glyptotek_colors}
\title{Extract Glyptotek colors as Hex Codes}
\usage{
glyptotek_colors(...)
}
\arguments{
\item{...}{Names of colors to extract.}
}
\value{
A character vector with color codes.
}
\description{
Can be used to get the hex code of specific colors from the Ny Calrsberg
Glyptotek color palette.
Use `glyptotek_colors()` to see all available colors.
}
\examples{
glyptotek_colors()
glyptotek_colors("glyptotekBlue", "yellow1")

}
