% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_paimio.R
\name{paimio_colors}
\alias{paimio_colors}
\title{Extract Paimio colors as Hex Codes}
\usage{
paimio_colors(...)
}
\arguments{
\item{...}{Names of colors to extract.}
}
\value{
A character vector with color codes.
}
\description{
Can be used to get the hex code of specific colors from the Paimio Sanatorium
color palette.
Use `paimio_colors()` to see all available colors.
}
\examples{
paimio_colors()
paimio_colors("paimioBlue", "yellow1")

}
