% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_glyptotek.R
\name{palette_glyptotek}
\alias{palette_glyptotek}
\title{Ny Carlsberg Glyptotek color palette}
\usage{
palette_glyptotek(palette = "full", reverse = FALSE, ...)
}
\arguments{
\item{palette}{A string specifying the palette name.}

\item{reverse}{Logical, whether to reverse the palette.}

\item{...}{Additional arguments passed to other functions.}
}
\description{
Ny Carlsberg Glyptotek color palette
}
\details{
This function is usually not called directly, but from within
[`scale_color_glyptotek()`][scale_color_glyptotek].
}
