% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_paimio.R
\name{palette_paimio}
\alias{palette_paimio}
\title{Paimio Sanatorium color palette}
\usage{
palette_paimio(palette = "full", reverse = FALSE, ...)
}
\arguments{
\item{palette}{A string specifying the palette name.}

\item{reverse}{Logical, whether to reverse the palette.}

\item{...}{Additional arguments passed to other functions.}
}
\description{
Paimio Sanatorium color palette
}
\details{
This function is usually not called directly, but from within
[`scale_color_paimio()`][scale_color_paimio].
}
