% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_glyptotek.R
\name{scale_color_glyptotek}
\alias{scale_color_glyptotek}
\alias{scale_color_glyptotek_d}
\alias{scale_color_glyptotek_c}
\alias{scale_colour_glyptotek}
\alias{scale_colour_glyptotek_c}
\alias{scale_colour_glyptotek_d}
\alias{scale_fill_glyptotek}
\alias{scale_fill_glyptotek_d}
\alias{scale_fill_glyptotek_c}
\title{Ny Calsberg Glyptotek Color Palette}
\usage{
scale_color_glyptotek(
  palette = "full",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_color_glyptotek_d(
  palette = "full",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_color_glyptotek_c(
  palette = "full",
  discrete = FALSE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_colour_glyptotek(
  palette = "full",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_colour_glyptotek_c(
  palette = "full",
  discrete = FALSE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_colour_glyptotek_d(
  palette = "full",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_fill_glyptotek(
  palette = "full",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "fill",
  ...
)

scale_fill_glyptotek_d(
  palette = "full",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "fill",
  ...
)

scale_fill_glyptotek_c(
  palette = "full",
  discrete = FALSE,
  reverse = FALSE,
  aesthetics = "fill",
  ...
)
}
\description{
This color palette is based on the colors from the Ny Calrsberg Glyptotek
}
