% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_paimio.R
\name{scale_color_paimio}
\alias{scale_color_paimio}
\alias{scale_color_paimio_d}
\alias{scale_color_paimio_c}
\alias{scale_colour_paimio}
\alias{scale_colour_paimio_c}
\alias{scale_colour_paimio_d}
\alias{scale_fill_paimio}
\alias{scale_fill_paimio_d}
\alias{scale_fill_paimio_c}
\title{Paimio Sanatorium Color Palette}
\usage{
scale_color_paimio(
  palette = "full",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_color_paimio_d(
  palette = "full",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_color_paimio_c(
  palette = "full",
  discrete = FALSE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_colour_paimio(
  palette = "full",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_colour_paimio_c(
  palette = "full",
  discrete = FALSE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_colour_paimio_d(
  palette = "full",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_fill_paimio(
  palette = "full",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "fill",
  ...
)

scale_fill_paimio_d(
  palette = "full",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "fill",
  ...
)

scale_fill_paimio_c(
  palette = "full",
  discrete = FALSE,
  reverse = FALSE,
  aesthetics = "fill",
  ...
)
}
\description{
This color palette is based on the colors from the Paimio Sanatorium (\url{https://paimiosanatorium.com/}]
by architect Alvar Aalto.
}
