% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sdlh.R
\name{theme_sdlh}
\alias{theme_sdlh}
\title{Custom theme}
\usage{
theme_sdlh()
}
\value{
A ggplot2 theme object
}
\description{
This function applies a custom ggplot2 theme with optimized settings.
}
\examples{
ggplot(pressure, aes(temperature,pressure))+
geom_point()+
theme_sdlh()

}
